Moving lsp definition request out of Defaults.
